% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_quadratic}
\alias{is_quadratic}
\title{Is a Penalty Quadratic?}
\usage{
is_quadratic(pen, ...)
}
\arguments{
\item{pen}{A \code{\link{penalty}} object.}

\item{...}{Passed to methods.}
}
\value{
A single logical.
}
\description{
\code{TRUE} only for \code{\link{quadratic_penalty}}, whose matrix, rank,
null basis and log pseudo-determinant the marginal-likelihood generics
expose.
}
\examples{
is_quadratic(quadratic_penalty(diag(2)))
is_quadratic(ridge_penalty())

}
