% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{penalty_gradient}
\alias{penalty_gradient}
\alias{penalty_hessian}
\title{Coefficient Derivatives of a Penalty}
\usage{
penalty_gradient(pen, beta, theta, ...)

penalty_hessian(pen, beta, theta, ...)
}
\arguments{
\item{pen}{A \code{\link{penalty}} object.}

\item{beta}{A numeric vector of coefficients.}

\item{theta}{A named list of hyperparameter values.}

\item{...}{Passed to methods.}
}
\value{
\code{penalty_gradient} a numeric vector of length \code{q};
  \code{penalty_hessian} a \code{q x q} symmetric matrix.
}
\description{
\code{penalty_gradient} returns \eqn{\partial\rho/\partial\beta} and
\code{penalty_hessian} returns \eqn{\partial^2\rho/\partial\beta^2}, both
exact.
}
\examples{
pen <- quadratic_penalty(diag(2))
penalty_gradient(pen, c(1, -1), list(lambda = 3))
penalty_hessian(pen, c(1, -1), list(lambda = 3))

}
