% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scad_mcp.R
\name{penalty_value.ScadPenalty}
\alias{penalty_value.ScadPenalty}
\alias{penalty_gradient.ScadPenalty}
\alias{penalty_hessian.ScadPenalty}
\alias{penalty_grad_theta.ScadPenalty}
\alias{penalty_hess_theta.ScadPenalty}
\alias{penalty_cross.ScadPenalty}
\alias{penalty_kinks.ScadPenalty}
\alias{is_proper.ScadPenalty}
\title{SCAD Methods}
\arguments{
\item{pen}{A \code{ScadPenalty} object.}

\item{beta}{A numeric vector of coefficients.}

\item{theta}{A list containing \code{lambda} and \code{a}.}

\item{scale}{Handled by the generic.}

\item{...}{Unused.}
}
\value{
See the generic pages.
}
\description{
The closed piecewise forms; see \code{\link{scad_penalty}}.
}
\keyword{internal}
